globalrevs: allow making commits without using pushrebase
Summary:
As an intermediate step in conversion of the `WWW` repository to being
an Mercurial only repository, we will be modifying the `hgsubversion` extension
to start supporting `globalrev` in the commits pulled from the source
Subversion repository.

The `globalrevs` extension needs to allow for creation of commit through means
other than pushrebase for this process to succeed. Therefore, let's introduce a
configuration option for this purpose.

Reviewed By: quark-zju

Differential Revision: D8925894

fbshipit-source-id: ecb253828c0801c1c39f57ccaf2e11992a924e63