acl: drop the extension
Summary:
It's rarely used here and feels like a burdern to continue support.

It only restricts write access. A well-designed commit hook API should be the
superior solution here.

The direct motivation of the removal is becuase it does not work well with the
Rust config parser. Namely it's the unique extension that treats an empty
section differently from an section that isn't set. That does not play well
with the Rust config parser. While it can be worked around like D8887670, it's
not a very clean soution, and there are other issues making test-acl.t fail.

Therefore, instead of trying to find the remaining weirdness of acl, I figured
it's probably the right call to drop it and redirect users to hooks instead.

Reviewed By: DurhamG

Differential Revision: D8953125

fbshipit-source-id: 209f04423ccdf182b0ba4021b59d26525173d73c