notify: drop the extension
Summary:
Similar to the acl extension, notify is not used here and can be implemented
as hooks.

The direct motivation is, it uses weird key in config names that get rejected
by the Rust config parser:

```
  hg: parse error: ".../.hgrc":
    --> 61:4
     |
  61 | */a#branch(test) = will_no_be_send@example.com
     |    ^---
     |
     = expected equal_sign
```

Although the pest syntax can be updated to support this case, by using
different `name`s for sections, and config names. It seems cleaner to
just drop notify, as it's the only trouble maker.

Reviewed By: DurhamG

Differential Revision: D8953127

fbshipit-source-id: fc7760a32b6ea538cb207c860aadac7b4b58027f