disable fastheaddiscovery during push
Summary:
fastheaddiscovery is a remotenames feature which makes discovery faster but
less precises. For example, it can request more commits from the server than
necessary. That's not a big deal during pulls, but it can be a problem during
push. It may try to push public commits to the server. People can ran into this
behaviour when they pull commit cloud commit from another devserver and this
commit is based on a newer public commit.

Reviewed By: quark-zju

Differential Revision: D8931322

fbshipit-source-id: 03c6f1df23c3742bda12adf40bf477711f7835bb