mononoke: make test less flakey
Summary:
The test was working fine, however, with the next diff in the stack
it started to fail under stress runs. And in that case I assume the problem in
the test itself.

The test creates two commits, that create file with the same content.
However, one filenode wasn't ancestor of another, and that's semantically
incorrect.

Reviewed By: farnz

Differential Revision: D8913129

fbshipit-source-id: 1b3838c6cf5060d768e71ef056cb7765ebbabc99