chg: check validate time instead of hello time for detecting slow servers
Summary:
There is an issue on OSX where the chg worker runs much slower. Its root cause
is still yet to be figured out (does OSX compresses background process memory
to make). It seems we hit it again according to https://fburl.com/co1eagws.

The previous workaround was to check "hello" time, and has a threshold to
restart the server. "hello" turned out to be a cheap operation and could be
inaccurate. From P59864987 it seems much more accurate to check the "validate"
operation (0.56s vs 0.026s). So this patch does that instead.

Note: we might need to revisit this if the Rust config parser lands, as that
will speed up the "validate" command.

Reviewed By: singhsrb

Differential Revision: D9019429

fbshipit-source-id: b2c02d08d20cc075d55e2d745d9394c90e543a1b