rage: show most recent 100 obsmarkers
Summary:
`hg rage` attempts to find interesting obsmarkers related to the user's
smartlog.  This turns out to be very slow for users with lots of obsmarkers,
and generates very long rage output in this case.

Instead, just put the last 100 obsmarkers in the rage output.  If earlier
obsmarkers are needed, we can look for them separately.

Reviewed By: ryanmce

Differential Revision: D9027952

fbshipit-source-id: 0368d5e83bced77bb863d223927d614407f19fe1