treemanifest: prevent pushrebase from processing trees during reply
Summary:
In a previous diff we stopped pushrebase from sending flat manifests
back to treeonly client, but we didn't stop pushrebase from analyzing from
analyzing the manifests to decide what files to send. It turns out that this can
be expensive, since it needs to diff two manifests.

Let's fix up the bundle2 capabilities so that pushrebase knows to not send files
back to the client for a remotefilelog client.

Reviewed By: markbt

Differential Revision: D9021066

fbshipit-source-id: ba422ce8b5ada7905e4900ec68957a84613897ff