tinyproxy: find free port automatically
Summary:
Previously tinyproxy.py was using `$HGPORT1` as its port, which might be
already taken by another process. If the port was taken, that could cause the
.t test to wait forever therefore the whole test has to time out.

This patch changes tinyproxy to use a free port assigned by the OS, write
the port down to a file so the test knows what port it is.

As we're here, clean up tinyproxy's argv handling so it relies on the option
parser, does not do hacky things on sys.argv, and only listen on localhost.

Reviewed By: DurhamG

Differential Revision: D8884144

fbshipit-source-id: 650ed3c637b136cbaee8280e1033edd0eb3ddc22