Refactor ReachabilityIndex tests to accept a closure instead of a trait object
Summary:
The tests for ReachabilityIndex now accept a closure that creates a trait object, instead of a trait object directly.
I was having problems later in this stack where I couldn't pass a SkiplistIndex object into these tests, because it help Arcs and hence wasnt 'UnwindSafe'. Modifying the tests to take a closure that creates the index object inside of the tokio async unit test fixes this.

Reviewed By: StanislavGlebik

Differential Revision: D9009053

fbshipit-source-id: de5e698b4d27aec1ab47fa7fda73da8ad46aab95