`fetch_generation_and_join` helper function, and refactoring of helper functions to use `impl Future`
Summary:
Added a helper function which attempts to fetch the Generation number for a node hash, and if successful, returns the (HgNodeHash, Generation) tuple. I found myself needing this in several places while working on the SkiplistIndex, and will probably continue to need it, so I gave it its own method.

Refactored existing helper methods to use `impl Future` instead of `BoxFuture`, since it was pointed out that this could be a performance issue in lower level helpers. This didn't change any functionality or require refactoring elsewhere, it just involved changing the signatures and removing any calls to `boxify()`.

Reviewed By: StanislavGlebik

Differential Revision: D9009057

fbshipit-source-id: 6b30bb92513213787fc92bdca1507ed117f6e715