Moving / deleting some struct elements in `SkiplistIndex` and `SkiplistEdgeMapping`
Summary:
* Moved the field `skip_edges_per_node` from `SkiplistIndex` to `SkiplistEdgeMapping`.
* Removed the `Arc<BlobRepo>` field from `SkiplistIndex`.

The first change was to avoid having to pass around the `skip_edges_per_node` value into all the closures. Since these closures include an `Arc<SkiplistEdgeMapping>`, they can just get the value from that.
The second change is to be more in line with the `ReachabilityIndex` trait that I defined and to fit more easily into how I wrote the tests which take a generic `ReachabilityIndex` object. One other possibility, if I decide that the object should hold onto a reference to the repo, and not receive it as input at query time, is to add another trait method `use_repo`, which would start using a given `Arc<BlobRepo>`. A single instance of `ReachabilityIndex` shouldn't be used for different repos, and this could be a way to check that at compiler time. But right now I'm assuming it's passed at query time.

Reviewed By: StanislavGlebik

Differential Revision: D9009059

fbshipit-source-id: 9d63acf8a690816b3a17c84108a6feaa98ebe2ba