bundle2: use repo bundle2 capabilities when making local bundles
Summary:
In a future diff we'll need to read a repository's bundle2 capabilities
while creating a strip backup bundle. To do so, we need to pass the bundle2
capabilities to the strip backup bundle creation logic.

This may enable a bunch of features in strip backup bundles now, like newer
formats that were gated behind capabilities.

Reviewed By: quark-zju

Differential Revision: D9075304

fbshipit-source-id: 17109284d2cd81ae467474089c1438c11edcd085