cloud join - add workspace parameter
Summary:
Add this as an experimental option

We don't have switching of workspaces so far, but hg clone is very fast with eden, so it is easy to create a brand new repo and join to any workspace.

This is already useful for debugging and repro user's issues.

You just join to the user's workspace from a brand new repo and can see what is going on there.

Reviewed By: quark-zju

Differential Revision: D9082157

fbshipit-source-id: 561314cddfced07c81c4bb6c1c2ccf99e140b27c