Implement reading of file contents in file hooks
Summary:
This diff gets the machinery in place to allow file hooks to access file content. It uses the manifest vfs crate to do this.

Initially we just expose a simple contains_string() method which checks if the file content contains the given string (this is the most common thing file hooks want to do with content), but we can extend this with further methods as the need arises.

Reviewed By: StanislavGlebik

Differential Revision: D8889837

fbshipit-source-id: fb5191d6d1286f070a99108782ff529914aeb198