commit cloud: fix cloud join command when no new heads
Summary:
We shouldn't send a get_bundles request with empty list of heads to the server.

It is waste of time for performance and also there is a bug to handle it on the server.

We build an incorrect mysql query.

I will fix it on server side as well.

Reviewed By: DurhamG, quark-zju

Differential Revision: D9184401

fbshipit-source-id: 24cf9db33f1faf5bf790f1d20766d250eb58b21b