Implement reading of file contents in file hooks for Lua hooks
Summary:
This diff extends Lua file hooks to allow them to access file content. Instead of exposing just the file path we now expose a Lua table which contains two fields:

'path' - the file path
'contains_string' - a function that can be called to see whether the file content contains the given string.

contains_string is executed as a coroutine which is immediately yielded so the hook does not block while waiting for the result (this potentially might involve network access).

Reviewed By: StanislavGlebik

Differential Revision: D8889838

fbshipit-source-id: 48b49e8ba2acb2d5479405f2ac27fbf5fc2d642a