lock: rename locker to lockinfo and clarify its usage
Summary:
The `locker` name is a bit confusing since it has nothing to do with actual locks.
It's just a read-only description.

Rename `locker` class to `lockinfo` and rename `lockerdesc` variable to `info`.
The `info` word comes from `util.makelock(info, path)`.

As we're here, also clean up the docstring.

Reviewed By: DurhamG

Differential Revision: D9004612

fbshipit-source-id: bc57cdd5d3196d0fb80de16a56c1cee0a8fd8486