configparser: strip leading space from multi-line value
Summary:
Multiline value like:

  [section]
  foo = a
    b

should be parsed as "a\nb", instead of "a \nb".

It does not affect configlist, but affects template definations.

Unfortunately in this case we had to allocate a new buffer instead of using
`Bytes::slice`. Fortunately most configs are single-line, so the performance
impact is hardly visible practically.

Reviewed By: DurhamG

Differential Revision: D8960866

fbshipit-source-id: 011e7f431d682236529ce176fe577aac6a010d91