configparser: translate byte range to line number at Python land
Summary:
The configsource line number is exposed in a bunch of UI and tests.  Therefore
translate the bytes range to line numbers in the Python wrapper.

The Rust config parser does not track line numbers explicitly, as that adds
complexity to the parser.

The implemenation is inefficient. However, configsource is not used in
common code paths, so it'd be fine.

Reviewed By: DurhamG

Differential Revision: D8960868

fbshipit-source-id: 55e9f5170177e7cb4c8ef55ed17c1f85eb70a192