fbshow: do not pass non-string config name to ui.setconfig
Summary:
The old code is passing tuples like `("B", "ignore-blank-lines", None, _("ignore changes whose lines are all blank"))`
as config name to `ui.setconfig`, which is clearly wrong.

Also, the CLI flag does not match configs in the `[diff]` config section, ex.
`diff.ignore-blank-lines` is not the config option. `diff.ignoreblanklines` is.
`diff.ignore-all-space` is not the config option. `diff.ignorews` is.

The diffopts config overrides are not necessary since the options are passed to
the diff command any way. So just drop it.

Reviewed By: DurhamG

Differential Revision: D8945812

fbshipit-source-id: 97eb2b1020c7ed8036a06a74f1bbb39bb545721b