uncommit: fix "restoring deletion" case for treestate
Summary:
The uncommit command does some low-level changes to dirstate - it changes the
parents without updating the tree. Then it uses traditional dirstate APIs to
update the tree.

Treestate and traditional dirstate differ on the "remove" API - treestate would
make it a no-op-ish if the file is not already tracked. Traditional dirstate
would create such an entry blindly and then mark it as "removed".

Treestate cannot simply create the entry because it needs "EXIST_P1" or
"EXIST_P2" information, which is outside the "remove" API itself.

To fix this, teach "uncommit" to write a treestate entry in this case.

The longer term fix would probably be removing the low-level "setparents"
calls.

Reviewed By: DurhamG

Differential Revision: D9282452

fbshipit-source-id: 9a0f0ac2f1bc2a4296c10f95819457f4f9585a82