historystore: add traversal options for ancestor traversals
Summary:
Previously ancestor traversals would return an error if they
encountered a node that couldn't be resolved. In some cases we want to support
partial ancestor resolutions (like iterating over part of a history in one
store, while the rest in another store). Let's add an option that decides
whether a missing node is an error or not.

Reviewed By: markbt

Differential Revision: D9231397

fbshipit-source-id: ff3063acfb8da2d453f34221f1865f3123615b0c