cloud sync: optimize cloud sync retries with validation on server side
Summary:
if there is a lot to push, it is worth to check with the server first

currently there is no any checks inside hg push implementation for commit cloud commits

hopefully it will be implemented in mononoke (heads and known wire proto request) and this logic will not be needed. Client and server will 'discuss' and figure out the delta what should be pushed and therefore any filtration in cloud sync will not be needed.

Reviewed By: quark-zju

Differential Revision: D9345385

fbshipit-source-id: 01a345905597af215dbf9b7609272491518abdf8