Rebalance cachelib less frequently
Summary:
Rebalancing causes cachelib to do early eviction of a slab in order to
increase hit rate in the cache as a whole. Now that we have much larger caches,
doing this infrequently makes sense - increase from every 10 seconds to every 5
minutes.

This will decrease the frequency of rebalancing logs

Reviewed By: StanislavGlebik

Differential Revision: D9343958

fbshipit-source-id: 42a33ac310083585318c1fc503bf0b1717b9fb61