globalrevs: allow specifying custom repository name for globalrevs
Summary:
We can have situations where using the repository name from the
`hgsql.reponame` configuration as key for `globalrevs` data in the database can
lead to issues. As an example, for some repo `x`, we usually have a
`x-from-svn` Mercurial repository which gets the commits from svn and then
repository `x` pulls that data from `x-from-svn`. In such a setup, the
`hgsql.reponame` is usually set to `x-from-svn` for the corresponding
repository and similarly for the repository `x`. For the purpose of getting rid
of the backing Subversion repository, we will need to use the `globalrevs`
extension in the `x-from-svn` repository and it makes sense for us to have the
repository name as `x` from the beginning instead of `x-from-svn` from a
`globalrevs` perspective.

This commit just allows specifying a custom repository name to help with
aforementioned situation.

Reviewed By: quark-zju

Differential Revision: D9372695

fbshipit-source-id: 080e08a4a3ab3a4ba7a213e5177e40deb5dd5d88