hg: distinguish binary from python script entry points
Summary:
To correctly create the hg command to shell out to, we need to know whether
we run from a binary or from a shebang script. If we run from a script,
the hg command should be `argv[0]` (as seen by Python), whereas when this is
done from a binary, we need to report the `sys.executable` (which Python pops
from `argv`).

Once I learn how to pass data down from Rust to Python, we can replace this
with something like a `rustdata` module and thus
```
try:
  import rustdata
  # we are called from Rust
except ImportError:
  # we are called from Python
```
will let us distinguish the two cases (given appropriate `hgdemandimport`
fixups).

Reviewed By: StanislavGlebik

Differential Revision: D9338851

fbshipit-source-id: d135bfd9a5d695e5d15fd79c5438758951656f0d