config: be compatible with Windows-style environment variables
Summary:
On Windows `%include` can have paths containing environment variables like
`%PROGRAMDATA%`. We already ship that kind of config files to users therefore
let's add support for that.

The change assumes `%` is not used as part of a normal path, which is probably
good enough for practical uses. If `%` does need to be legally used in a
filename, we can add escaping support later.

Reviewed By: DurhamG

Differential Revision: D9283303

fbshipit-source-id: bcc80307fe19dfc40aea88b6a0a5f69681e835fc