checklink and checkexec avoid mutating eden filesystem root inode
Summary:
While examining a system where Eden was under load we discovered
that an application had gotten into a feedback loop; it was asking watchman
to tell it about changes in the repo, and each time a change occurred it
would run `hg status` and `hg log`.  These commands triggered the checkexec
and checklink code to modify the filesystem, which in turn generated change
events for the application to consume.

The application already filters out changes under the `.hg` directory, but
these probe files are not placed there.  Inspecting the code shows that it
has the ability to use `.hg/cache` for this purpose, but as far as I can
tell, nothing ever creates this directory in the working copies in common
use.

I made an initial pass at creating that directory in these functions in
D9322798, but there are hg tests that are sensitive to the precise mode of that
dir.  The semantics and creation of that dir are not clear, so I'm just going
for a simpler option: moving the files under the `.hg` dir if that dir exists.

Moving these files is desirable because it simplifies ignore logic for
downstream clients, and takes pressure off the root inode in an eden mount,
which is more expensive to save changes into than the others due to its
stronger durability requirements.

However, the `.hg` directory may not even be on the same filesystem as the
rest of the working copy.

For the sake of an expedient fix, this diff checks to see if we're inside an
eden mount and uses that information to inform the caller; eden mounts
on posix systems support both symlinks and exec bits.

Reviewed By: quark-zju

Differential Revision: D9328199

fbshipit-source-id: 38af8e3d95e7acee103a1fcc52277511fd45e9fc